requirements.txt
fastapi>=0.68.0,<0.69.0
pydantic>=1.8.0,<2.0.0
uvicorn>=0.15.0,<0.16.0
anyio==3.5.0
asgiref==3.5.0
click==8.1.2
colorama==0.4.4
cycler==0.11.0
Flask==2.1.1
fonttools==4.31.2
h11==0.13.0
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.1
kiwisolver==1.4.2
MarkupSafe==2.1.1
matplotlib==3.5.1
numpy==1.22.3
packaging==21.3
pandas==1.4.2
Pillow==9.1.0
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2022.1
six==1.16.0
sniffio==1.2.0
starlette==0.14.2
typing_extensions==4.1.1
Werkzeug==2.1.1